requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fschat==0.2.32
fsspec==2023.10.0
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nh3==0.2.14
numpy==1.26.1
openai==1.1.2
orjson==3.9.10
packaging==23.2
pandas==2.1.2
peft==0.6.0
Pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pydantic==1.10.13
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
semantic-version==2.10.0
sentencepiece==0.1.99
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tiktoken==0.5.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
wavedrom==2.0.3.post3
wcwidth==0.2.9
websockets==11.0.3
xopen==1.8.0
yarl==1.9.2
